week-8/03-Streamlit/requirements.txt
folium==0.12.1.post1
pandas==1.1.4
Pillow==9.3.0
plotly==5.10.0
requests==2.26.0
streamlit==1.12.0
streamlit_folium==0.6.14
